feat:uni-电子签名

+++ docs/.vitepress/script/LastTime.js
@@ -3,7 +3,7 @@ const lastUpdated = {
   text: '更新时间',
   formatOptions: {
     locale: 'zh-CN',
-    dateStyle: 'long',
+    dateStyle: 'short',
     // dateSty用于指定日期的样式，可选值包括 'full'、'long'、'medium' 和 'short'
     timeStyle: 'medium',
     // timeStyle用于指定时间的样式，可选值包括 'full'、'long'、'medium' 和 'short'
